+++ Project/app/src/main/java/com/example/quanlynhahang/retrofit/HoaDonAPI.java
@@ -9,6 +9,6 @@ import retrofit2.http.Query;
 public interface HoaDonAPI {
     @POST("hoadon/tao")
     Single<HoaDon> taoHoaDon(@Query("maDonHang") int maDonHang, @Query("maNhanVien") int maNhanVien, @Query("maNhaHang") int maNhaHang,
-                             @Query("tongTien") double tongTien,
+                             @Query("tongTien") int tongTien,
                              @Query("phuongThuc") String phuongThuc);
 }
